pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "lerobot_robot_rerun"
version = "0.1.0"
description = "Generic URDF robot visualizer for LeRobot — streams joint states to a rerun viewer"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "Koen van Wijk" },
]
keywords = ["lerobot", "rerun", "urdf", "visualization", "robotics"]
dependencies = [
    "lerobot>=0.5.0",
    "rerun-sdk>=0.16",
    "yourdfpy>=0.0.56",
    "numpy>=1.24",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."lerobot.robots"]
rerun_robot = "lerobot_robot_rerun:RerunRobotConfig"
